requirements.txt
flask>=2.0
google-generativeai>=0.3.0
python-dotenv>=1.0
